### 2019-07-26 #1600 @whedon
[PRE REVIEW]: DscoreApp: An user-friendly web application for computing the Implicit Association   Test D-score
**Submitting author:** @OttaviaE (<a href="http://orcid.org/0000-0001-8552-568X">Ottavia M. Epifania</a>)
**Repository:** <a href="https://github.com/OttaviaE/DscoreApp" target ="_blank">https://github.com/OttaviaE/DscoreApp</a>
**Version:** v1.0.0
