### 2019-07-26 #1600 @kyleniemeyer
Hi @OttaviaE, thanks for your submission to JOSS!

Unfortunately, we currently only support papers written in Markdown; our documentation has a good example that shows all the features supported: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography

Can you please convert your current `paper.Rmd` to the format we support, and let us know when you have? Thanks!
