### 2019-07-26 #1600 @kyleniemeyer
@OttaviaE no worries! But actually, I think the converted Markdown file is still not quite right. We do use the YAML header like what you had before, cite references from the `.bib` file, and support Markdown tables and figures.

I will try to do a quick pass through and fix it up for you.
