### 2019-07-26 802fb7e
paper JOSS md/YAML header
