+++ paper.md
@@ -1,255 +1,93 @@
-Summary
-=======
+---
+title: 'DscoreApp: An user-friendly web application for computing the Implicit Association
+  Test D-score'
+authors:
+- affiliation: '1'
+  name: Ottavia M. Epifania
+  orcid: 0000-0001-8552-568
+- affiliation: '1'
+  name: Pasquale Anselmi
+  orcid:  0000-0003-2982-7178
+- affiliation: '1'
+  name: Egidio Robusto
+date: "26 luglio, 2019"
+output: 
+  html_document:
+       keep_md: true
+  pdf_document: default
+bibliography: paper.bib
+tags:
+- R
+- Shiny
+- Implicit Association Test
+- D-score
+- User-friendly
+affiliations:
+- index: 1
+  name: Department of Philosophy, Sociology, Pedagogy and Applied Psychology, University
+    of Padova (IT)
+---
 
-The Implicit Association Test (IAT; Greenwald, McGhee, and Schwartz
-1998) is one of the most commonly used measures for the implicit
-assessment of attitudes and preferences. It is based on the speed and
-accuracy with which stimuli representing four different categories
-(e.g., flowers and insects, positive and negative words in a
-Flowers-Insects IAT) are sorted in their belonging category by means of
-two response keys. The IAT is usually composed of seven blocks, three of
-which are pure practice blocks where either flowers-insects images or
-positive-negative words are sorted in their categories. The remaining
-four blocks are associative practice and associative test blocks. These
-blocks forms the two contrasting associative conditions under which the
-categorization task takes place. In one associative condition (i.e.,
-practice blocks Mapping A and test blocks Mapping A), flowers images and
-positive words share the same response key, while insects images and
-negative words share the other response key. In the contrasting
-associative condition (i.e., practice blocks Mapping B and test blocks
-Mapping B), insects images and positive words share the same response
-key, while flowers images and negative words share the opposite response
-key.
 
-The categorization task is supposed to be easier (i.e., faster response
-times and higher accuracy) in the associative condition consistent with
-respondents’ automatically activated associations. The *IAT effect*
-results from the difference in respondents’ performance between the two
-contrasting conditions, and the *D-score* (Greenwald, Nosek, and Banaji
-2003) is the most common measure used for interpreting the strength and
-direction of this effect. Despite different options are available for
-the *D-score* computation (Table 1), the core procedure is the same. The
-difference between the *D-score*s only concerns the treatment for the
-error responses and the treatment for the fast responses.
+# Summary
 
-<table>
-<caption><em>D-score</em> algorithms overview. Trials with latency <span class="math inline">&gt;10,000</span> ms are discarded for all the algorithms.</caption>
-<thead>
-<tr class="header">
-<th style="text-align: left;">D-score</th>
-<th style="text-align: left;">Error treatment</th>
-<th style="text-align: left;">Lower tail treatment</th>
-</tr>
-</thead>
-<tbody>
-<tr class="odd">
-<td style="text-align: left;">D1</td>
-<td style="text-align: left;">Built-in correction</td>
-<td style="text-align: left;">No</td>
-</tr>
-<tr class="even">
-<td style="text-align: left;">D2</td>
-<td style="text-align: left;">Built-in correction</td>
-<td style="text-align: left;">Delete trials <span class="math inline">&lt;</span> 400 ms</td>
-</tr>
-<tr class="odd">
-<td style="text-align: left;">D3</td>
-<td style="text-align: left;">Mean (correct responses) <span class="math inline">+2</span> <em>sd</em></td>
-<td style="text-align: left;">No</td>
-</tr>
-<tr class="even">
[…]
