### 2019-07-26 #1600 @OttaviaE
@kyleniemeyer  This is quite odd because my ORCID ends with an "X"
