### 2019-07-26 #1600 @kyleniemeyer
🎉 !

It looks like there are still some issues with the figures, which I should be able to fix quickly. One minute.
