### 2019-07-26 #1600 @kyleniemeyer
@OttaviaE OK, please merge https://github.com/OttaviaE/DscoreApp/pull/1, that should do the trick.
