### 2019-07-26 #1600 @kyleniemeyer
OK @alexhanna, all yours!
