### 2019-09-05 #1600 @benmarwick
I reviewed this submission using the [guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) and [checklist](https://joss.readthedocs.io/en/latest/review_checklist.html). Items needing further attention by the authors are marked with a 👀Disclaimer: I'm not a psychologist, don't have a strong understanding of what a D-score is, so my review does not evaluate the correctness of calculations performed by this app.

Looking at the [paper.pdf](https://github.com/openjournals/joss-papers/blob/joss.01600/joss.01600/10.21105.joss.01600.pdf), here's what I saw:

- [x] A list of the authors of the software
- [x] Author affiliations
- [ ] A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience -> 👀 I recommend the authors identify the research community that will find this useful, e.g. "...most commonly used measures in psychology for ..."
- [x] A clear statement of need that illustrates the purpose of the software
Mentions (if applicable) of any ongoing research projects using the software or recent scholarly publications enabled by it
- [x] A list of key references including a link to the software archive

- [x] As the reviewer I confirm that I have read the JOSS conflict of interest policy and that there are no conflicts of interest for me to review this work.

- [x] I confirm that I read and will adhere to the JOSS code of conduct.

- [x] Repository: Is the source code for this software available at the repository url?
- [x] License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?
- [ ] Version: Does the release version given match the GitHub release (v0.8)? -> 👀 I could not find version numbers, these need to be made easier to see
- [x] Authorship: Has the submitting author made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

- [ ] Functionality: Have the functional claims of the software been confirmed? -> 👀 I had a problem running the app because of a absolute path in `server.R` for the example data. After I changed L22 to `dataset <- read.csv(here::here("raceAPP.csv"))` then the data could load in the shiny app and everything worked as expected.

- [ ] Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified? -> 👀 I didn't see these
- [ ] Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support  -> 👀 I didn't see a CONTRIBUTING.md or similar in the app or on the repo.

- [x] Authors: Does the paper.md file include a list of authors with their affiliations?
- [x] References: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?
