### 2019-09-06 c77f950
Adding version of the app and identify the research community for the app
