### 2019-09-06 7e93100
Adding explanation on how to use the example dataframe locally.
