+++ README.md
@@ -1,7 +1,13 @@
-# DscoreApp
+# DscoreApp (v0.1)
 
 A Shiny Web Application for computing the IAT D-score. 
 
-This app is meant as an easy and open source alternative for computing the IAT D-score.
+This app is meant as an easy and open source alternative for computing the IAT 
+D-score.
 
 DscoreApp can also be used online, and it is retriavable at the URL: http://fisppa.psy.unipd.it/DscoreApp/
+
+If you want to use the app locally on your computer, you have to set the 
+directory on LINE 23 in `server.R` to your working directory (if you want to use
+the example dataset). 
+
