### 2019-09-06 #1600 @OttaviaE
Hey @benmarwick thank you for your review! I tried to solve the issues you highlighted:
1.	“I recommend the authors identify the research community that will find this useful, e.g. "...most commonly used measures in psychology for ...". Done
2.	“I could not find version numbers, these need to be made easier to see”. I add the version number in the README.md file
3.	“I had a problem running the app because of a absolute path in `server.R` for the example data.”. Since the app is meant for being used online at the URL http://fisppa.psy.unipd.it/DscoreApp/, I added the instructions for changing the wd for the example dataset in the README.md file.
4.	“Automated tests”. You can find the tests for the app functioning here https://github.com/OttaviaE/DscoreApp/tree/master/tests plus a `running_test.R` script in the GitHub directory
5.	“I didn't see a CONTRIBUTING.md or similar in the app or on the repo.” I added the CONTRIBUTING.md file to the repo. On the online version of the app at http://fisppa.psy.unipd.it/DscoreApp/ I left my personal email for reporting bugs/ask for support under the “Contacts” in the “Read Me First” panel.

Hope it helps!
