### 2019-09-06 #1600 @arfon
Thanks for your review here @benmarwick and your response @OttaviaE. I realize it might not be entirely clear but the actual JOSS review happens in a separate issue on GitHub that is set up by the editor (@alexhanna) once they have identified multiple reviewers.

I believe @alexhanna still needs to identify a second reviewer for this submission and will set up the main review issue at that point.

@benmarwick - we can just copy over your review comments at that point :-)
