### 2019-09-19 #1600 @OttaviaE
@russellpierce it's not actually on the IAT itself, it's more focused on the score as it's usually computed according to Greenwald et al. (2003).
