### 2019-09-19 #1600 @OttaviaE
@alexhanna Since @russellpierce is not suitable for this review, maybe @sctyner would like to do it
