### 2019-09-25 #1764 @tomfaulkenberry
I have completed my review of DscoreApp and the associated software paper.  DscoreApp is a very nice, fully-featured Shiny app for computing D-scores from Implicit Association Test data.  I was able to follow the instructions provided by the authors for the example data.

My main comments are related to the software paper itself, not the code.  Thus, instead of opening an issue on the Github repo for DscoreApp, I'll just list them here.  These are primarily editorial comments that I think will make the reading of the paper easier.

1. In paragraph 1, the phrases "flowers images" and "insects images" are used frequently.  It may be more clear to say instead "images of flowers" and "images of insects".

2. Underneath Table 1, I suggest editing "respondents might be given a feedback" to "respondents might be given feedback" (i.e., remove the "a" from the sentence)

3. For the three-step procedure on computing D-scores:  can you explain exactly what standard deviation is being computed?  From what is written, I'm not sure whether the standard deviations of each block are being pooled, or whether the trials are all pooled and then standard deviation is computed.

4. In the section DscoreApp:  The figures "D-score results panel" and "Results graphical representations" are not appearing in the paper (but they do in the version of the paper on the Github repo).  I'm assuming this is a minor issue caused by some glitch in the paper build, so it should be easy to fix.

Finally, I'll mention that there's no clear instruction in the paper or the Github repo about how to contribute to the project.  Given that this is open source software, some indication for how others can go about adding to the project is absolutely necessary (my suggestion -- just give a paragraph on the Github readme about how issues or requests can be opened).

Overall, I am quite happy with this package (actually, I'm very impressed with the coding!).  I think it will be a useful app for many people who use IATs.  Once my concerns noted above have been addressed, I will enthusiastically endorse publication in JOSS.
