### 2019-09-25 ddf4fd8
How to contribute to DscoreApp
