+++ CONTRIBUTING.md
@@ -1,13 +1,11 @@
-# DscoreApp (v0.1)
+# Contributing to DscoreApp
 
-A Shiny Web Application for computing the IAT D-score. 
+If you want to contribute to this app, you can open a new branch on https://github.com/OttaviaE/DscoreApp, modify the code, and submit your pull request for added features. 
 
-This app is meant as an easy and open source alternative for computing the IAT 
-D-score.
+To report any bug or any issue related to the app functioning, you can open a new
+issue on https://github.com/OttaviaE/DscoreApp/issues or send me an email at
+otta.epifania@gmail.com.
 
-DscoreApp can also be used online, and it is retriavable at the URL: http://fisppa.psy.unipd.it/DscoreApp/
+If you need support or you seek help, just contact me at otta.epifania@gmail.com.
 
-If you want to use the app locally on your computer, you have to set the 
-directory on LINE 23 in `server.R` to your working directory (if you want to use
-the example dataset). 
 
