+++ paper.Rmd
@@ -4,17 +4,17 @@ title: 'DscoreApp: An user-friendly web application for computing the Implicit A
 authors:
 - affiliation: '1'
   name: Ottavia M. Epifania
-  orcid: 0000-0001-8552-568
+  orcid: 0000-0001-8552-568X
 - affiliation: '1'
   name: Pasquale Anselmi
   orcid:  0000-0003-2982-7178
 - affiliation: '1'
   name: Egidio Robusto
 date: "`r format(Sys.time(), '%d %B, %Y')`"
-output:
-  pdf_document: default
+output: 
   html_document:
-    df_print: paged
+       keep_md: true
+  pdf_document: default
 bibliography: paper.bib
 tags:
 - R
@@ -31,7 +31,7 @@ affiliations:
 
 # Summary
 
-The Implicit Association Test [IAT; @Greenwald:1998] is one of the most commonly used measures for the implicit assessment of attitudes and preferences. It is based on the speed and accuracy with which stimuli representing four different categories (e.g., flowers and insects, positive and negative words in a Flowers-Insects IAT) are sorted in their belonging category by means of two response keys. The IAT is usually composed of seven blocks, three of which are pure practice blocks where either flowers-insects images or positive-negative words are sorted in their categories. The remaining four blocks are associative practice and associative test blocks. These blocks forms the two contrasting associative conditions under which the categorization task takes place. In one associative condition (i.e., practice blocks Mapping A and test blocks Mapping A), flowers images and positive words share the same response key, while insects images and negative words share the other response key. In the contrasting associative condition (i.e., practice blocks Mapping B and test blocks Mapping B), insects images and positive words share the same response key, while flowers images and negative words share the opposite response key. 
+The Implicit Association Test [IAT; @Greenwald:1998] is one of the most commonly used measures in psychology for the implicit assessment of attitudes and preferences. It is based on the speed and accuracy with which stimuli representing four different categories (e.g., flowers and insects, positive and negative words in a Flowers-Insects IAT) are sorted in their belonging category by means of two response keys. The IAT is usually composed of seven blocks, three of which are pure practice blocks where either flowers-insects images or positive-negative words are sorted in their categories. The remaining four blocks are associative practice and associative test blocks. These blocks forms the two contrasting associative conditions under which the categorization task takes place. In one associative condition (i.e., practice blocks Mapping A and test blocks Mapping A), images of flowers and positive words share the same response key, while images of insects and negative words share the other response key. In the contrasting associative condition (i.e., practice blocks Mapping B and test blocks Mapping B), images of insects and positive words share the same response key, while images of flowers and negative words share the opposite response key. 
 
 The categorization task is supposed to be easier (i.e., faster response times and higher accuracy) in the associative condition consistent with respondents' automatically activated associations. The *IAT effect* results from the difference in respondents' performance between the two contrasting conditions, and the *D-score* [@Greenwald2003] is the most common measure used for interpreting the strength and direction of this effect. Despite different options are available for the *D-score* computation (Table 1), the core procedure is the same. The difference between the *D-score*s only concerns the treatment for the error responses and the treatment for the fast responses. 
 
@@ -48,11 +48,11 @@ Table: *D-score* algorithms overview. Trials with latency $> 10,000$ ms are disc
 | D6      | Mean (correct responses) $+ 600$ ms     | Delete trials $<$ 400 ms | 
 
 
[…]
