### 2019-09-25 5f1aca9
Merge "README" with "CONTRIBUTING"
