+++ CONTRIBUTING.md
@@ -1,11 +1,0 @@
-# Contributing to DscoreApp
-
-If you want to contribute to this app, you can open a new branch on https://github.com/OttaviaE/DscoreApp, modify the code, and submit your pull request for added features. 
-
-To report any bug or any issue related to the app functioning, you can open a new
-issue on https://github.com/OttaviaE/DscoreApp/issues or send me an email at
-otta.epifania@gmail.com.
-
-If you need support or you seek help, just contact me at otta.epifania@gmail.com.
-
-
