### 2019-09-25 #1764 @OttaviaE
@tomfaulkenberry Thanks for your comments!

> _1.	In paragraph 1, the phrases "flowers images" and "insects images" are used frequently. It may be more clear to say instead "images of flowers" and "images of insects"._

Done.

> 2.	_Underneath Table 1, I suggest editing "respondents might be given a feedback" to "respondents might be given feedback" (i.e., remove the "a" from the sentence)_

Done.

> 3.	_For the three-step procedure on computing D-scores: can you explain exactly what standard deviation is being computed? From what is written, I'm not sure whether the standard deviations of each block are being pooled, or whether the trials are all pooled and then standard deviation is computed._

I changed the sentence to “This difference is divided by the standard deviation computed on the pooled trials of both blocks.”. I hope it’s clearer that the trials of the blocks are first pooled together, and then standard deviation is computed

>4.	_In the section DscoreApp: The figures "D-score results panel" and "Results graphical representations" are not appearing in the paper (but they do in the version of the paper on the Github repo). I'm assuming this is a minor issue caused by some glitch in the paper build, so it should be easy to fix._

I think it is a glitch, and last time @kyleniemeyer managed to fix it.

I already addressed the issues highlighted by @benmarwick, but I can copy and paste the reply here.
