+++ README.md
@@ -22,3 +22,6 @@ otta.epifania@gmail.com.
 
 If you need support or you seek help, just contact me at otta.epifania@gmail.com.
 
+## Acknowledgments
+
+O.M.E. wanted to thank Ben Keller for his suggestions and advices.  
