### 2019-09-25 #1764 @kyleniemeyer
@OttaviaE it looks like your figures are currently included using HTML commands, but the paper needs to use the Markdown commands instead.

So, for example, replace
```HTML
<p class="caption">D-score results panel.</p>
```

with

```markdown
![D-score results panel.](results.png)
```

(but for both figures)
