### 2019-09-25 #1764 @tomfaulkenberry
I'll attempt to regenerate the PDF here...
