### 2019-09-25 #1764 @tomfaulkenberry
New pdf looks great.  My review is complete...please consider this my final endorsement.  @OttaviaE, nice work indeed :)
