### 2019-09-26 #1764 @benmarwick
My review is here on the pre-review issue: https://github.com/openjournals/joss-reviews/issues/1600#issuecomment-528479908

I see responses by @OttaviaE here: https://github.com/OttaviaE/DscoreApp/tree/master/tests, thanks!

However, I don't see the high-level description anywhere on the repo or the app, can you point me to it more specifically? I'm not familiar with how the json files in the tests directory perform tests of the app, so I can't evaluate those (I use testthat). How do we know the tests are passing? Can the app repo be hooked to Travis so we can get a green badge to indicate that all tests are passing? I don't see CONTRIBUTING.md in the repo anyway, perhaps another git push will make it show?

I prefer to see shiny apps as packages because I can more easily identify where the core parts are, and understand the test coverage more readily. Then you can have a regular DESCRIPTION file to hold the version number, license details, and other metadata in a standardised, machine-readable way that most users expect to see.
