### 2019-09-26 #1764 @OttaviaE
@benmarwick thank for your reply.

By "high-level description" you mean like the DESCRIPTION file in the R packages? If so, I don’t whether a file like that would be useful/feasible for the app.

I added a "Contributing" section to the README.md of the app, in which I give instructions of how to contribute to the app and how to contact me in case of need or for signaling bugs.

Since a Shiny App is not an R package, you can't use testthat even if you want (I tried). I tested the app by using the shinytest package (https://cran.r-project.org/web/packages/shinytest/index.html). It basically runs the app a first time and save the expected outcomes, then runs a test in which it compares the first outcome with the second one. If there are some differences, it returns an error.
I developed this app with the idea of giving an easy-to-use tool for computing the D-score for the IAT. Indeed, the app is supposed to be used online at the link I posted in the README file, and I have created the GitHub repository so that people can actually check what the code if they want to do that, but it is not mandatory to understand how the app is working.
