### 2019-10-27 #1764 @alexhanna
I've checked off things in @benmarwick's review which were checked in issue #1600.

@OttaviaE - there doesn't need to be a DESCRIPTION, but according to [JOSS guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#documentation), the high-level description can be in the README.md file and include:

A statement of need
Installation instructions
Example usage
API documentation
