### 2019-10-28 #1764 @OttaviaE
Hey @alexhanna  I've just pushed some changes on GitHub that should addressed the issues in @benmarwick reviews.

Let me know whether it is okay!
