### 2019-10-29 #1764 @benmarwick
Thanks, sorry for the delay, on reflection I see I've been reviewing using the rOpenSci package review guidelines, which are not necessary here because the author is not submitting for onboarding with rOpenSci (unlike the last pkg I reviewed for JOSS). So, I'm ok with it, nothing further from me. 👍
