### 2019-10-29 #1764 @alexhanna
@OttaviaE the README looks better. @benmarwick, can you go through the checkboxes in the review and ensure that the software meets these criteria?
