### 2019-10-30 #1764 @OttaviaE
@danielskatz I updated the zenodo metadata according to your instructions. Concerning the version, it should be the v0.1.
