### 2019-10-30 588bc19
text changes
