### 2019-10-30 #1764 @danielskatz
Once the DOI resolves, I will close this issue, but it isn't yet working for me
